res.statusCode = 500 for default error handler (#28)
* res.statusCode = 500 for default error handler

node sets res.statusCode to 200 by default so 500 was never set even in
the default error handling case

* res.statusCode should reflect error in handler

res.statusCode is automatically set to 200, but in an error it should be set to 500. If res.statusCode has been already set to an error HTTP code (> 400), leave it be. err.statusCode takes precedence